fix(native): RCTRootView copies background color, workaround with NSNotificationCenter listening
Workaround for white flash between LaunchImage and JS splash screen was to set
backgroundColor with a pattern image, however RCTRootView copies the background
color from rootView when JavaScript finishes loading, which caused the LaunchImage
to be sometimes visible when transitioning between views (specially when using gestures to
navigate back), and also when rotating the device. This commit restores the initial background
when RCTRootView emits the notification that the bundle has finished loading.
This also addresses reloads where the launch image stays.