fix(library): allow scrolling
Allow scrolling on Library section. This was originally disabled
since we have an horizontally-scrolled Article list and we were
concerned on how the two would play together. There seem to be
no issues, so we're allowing scrolling (which is really just
bouncing on test devices) to be consistent with the rest of the tabs.