fix(stimulation): switch to refetchQueries after activity actions
updateQueries was missing cursor, which we fixed. But then we switched to refetchQueries since there's a part of the Profile query (baby's week ahead) that might potentially be displaying the activity that gets changed as a result of this mutation. Since we don't have enough data, and the logic to figure out whether the old activity needs updating, we're keeping it simple for now and just refetching the queries of the places were an activity action might affect, namely This Week's Activities and Profile.

ref: #108