perf: small improvement to loading speed
Remove the fetching of all babies on load for the Choose Baby, this should
technically be in case after the first time the user hits Choose Baby, if
anything we're going to make improvements there instead.
Moves quote processing to the server, and add new field that returns a random
quote. This would prevent React from having to compute that many diffs.

ref: #100