fix(growth): display abbrev months in graph detail
We reverted this at some point, but it's clear with new data
that we need to display abbreviated months in the horizontal
axis of graph detail, otherwise they overlap each other, and
the only viable option is to omit or ellipsis some. I think
abbreviated months should be fairly expected from users in a
graph like this. We could consider displaying full for months
below 5 or even 6 chars, but then it's not that clear.