build(native): workaround BB running XCode tests
* Link RNSVG to test project.
* Comment out generated test.

This is a workaround until we find a way to disable
BuddyBuild from running these kind of tests. We're not
doing E2E and this certainly slows our builds; something that
started happening just recently.

In the future, if we decide to indeed do E2E/iOS testing it will most likely
be not through XCTest but some of the (albeit limited) higher-level/more-RN-friendly
solutions.