feat: rewritte Choose Baby animation and improve interactions
* Improves animation smoothness significantly.
* Adds transition when closing it. Previously, it had no transition and might have
even contained some synchronous code in library code. This transition happens
smoothly when closing, and the only problem is triggering a barely noticeable
flash when going to Add Baby, since we need to reset the navigation stack
so that hitting the Back button takes you home instead of to the modal.
* Switch to `PureComponent` where applicable.
* Remove all the unused transitioners and change custom transitioner API to
avoid having to render a different component, which was an obvious bottleneck.
* Remove state and use of context from `TransitionerSwitcher`, while I don't like
binding vars to instance (we need to store previousTransition since react-navigation
fails to properly return an interpolator when coming back from a custom one), it's
significantly faster than passing context and calling `setState` (which triggers
a re-render).
* In order to improve responsiveness, we no longer destroy the header on opening
Choose baby, this has the implication of a small space to the top that I'm experimenting
with either child view positioning, or making use of it (i.e show currently selected
Baby name or "Choose a Baby" and moving close button up; designs are not that
different).
* Our work in landscape had an SVG path creation on the render method, this
was significantly slowing render cycles.

NOTE: adding the previously removed prefetched query might result in even better
responsiveness at the cost of app loading time. It looks good to me as is,
but after we land query splitting support this might be an option to consider, as
we wouldn't be sacrificing initial load time and the query would be fetching on
the background.