fix(schema): deprecate {created,changed} in payloads, use edge instead
Deprecate payload fields `createdBaby` and `changedType` in mutation
payloads. This ensures consistency with the rest of the app, and we
should `edge` instead. The reason is that the deprecated fields
don't provide enough information for manipulating lists and pagination,
or optimistic responses, and as a result we were using `refetchQueries` since
we had no `cursor` information. This change ensures we return all data we need
and we'll move away from query refetching in the future.