fix(profile): remove need to `refetchQueries` on baby update
This removes the need to use `refetchQueries` after updating a
Baby. The reasons we're able to make this happen now are:

1. Our uploads now return different filenames when you upload
a new avatar or cover image, so there's no need to refetch `getBabyAvatar`.
Previously, it returned the same URL so there was no way to tell connected
components to re-render other than refetching their queries.
2. Work on the previous commit to ensure we return all
data we need.