feat(profile): delete baby
Mark a baby for deletion, effectively making it unavailable to the
app. This adds a `wasDeleted` boolean to the baby that functions
can hook into. This field is now used to filter out the results
in `getBabies`.

Includes a retention period confirmation prompt, which could be implemented by
using a pair of functions, one that enqueues the deletion, and one that
actually does the cleanup.  See linked issue for the details of this required
cleanup.

ref: #123