feat: remove file upload from server and `FileInputBase64`
Removes file upload from server. It has always worked sub-optimally,
and it now depends on the client to do uploads, this ensures we can tap into
native optimizations of such process, remove brittle workarounds that were put
in place to accommodate for React Native's handling of file input (even with
`react-native-fetch-blob` and native `FormData` polyfills), and allow web
to do uploads.  We used to use a hijacked `FileInput` url to provide a (again,
polyfilled) `Blob` instance, which is not possible on server GraphQL.
Furthermore, I'm avoiding upload code on functions due to inconsistencies in
both GCF and AWS Lamda, plus the fact that both have timeouts and we want
to both avoid missing or incomplete uploads, or implementing a complete async
architecture (with @defer and queues) to make this happen. The simpler solution
is to let the client do the upload, which is almost complete on native, and
shouldn't have problems on web (tentatively, it should now work out of the box,
given how we've extracted injecting firebase and the file-aware code in
`ServerNetworkInterface`). As such, file upload is now handled at the interface-level,
meaning no code needs to change from a view perspective (other than perhaps sending
metadata differently, which wasn't implemented yet).

BREAKING CHANGE: Server does no longer do file uploads. We expect file fields
in mutations to include an already-working `url` in Firebase Storage. In
addition, `FileInputBase64` has been removed to better reflect that we were
using `FileInput` with an `url`.