feat: prepare core for native Firebase SDK
Due to previously-requested reasons, in addition to upload and performance,
offline support, and native thread offloading, we will be switching to native
Firebase SDK. This commit prepares the path for integrating
`react-native-firebase` (which we're already using for analytics and push
notifications) by allowing the caller to specify the Firebase instance. This
also removes the need of including Firebase twice due to features not supported
by the Web SDK.

BREAKING CHANGE: `configureDeps` now expects `firebase` to be passed through
`platformDeps`. Firebase should now be initialized by the caller of
`configureStore`, and we no longer import Web SDK by default.